+++ app/components/obs/inputs/GenericFormGroups.vue.ts
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import GenericForm from './GenericForm.vue';
 import { Component, Prop } from 'vue-property-decorator';
-import { ISettingsSubCategory } from '../../../services/settings/index';
+import { ISettingsSubCategory } from '../../../services/settings';
 
 @Component({
   components: { GenericForm },
