+++ app/components/page-components/Chatbot/windows/ChatbotDefaultCommandWindow.vue.ts
@@ -8,7 +8,7 @@ import { metadata as metadataHelper } from 'components/widgets/inputs';
 import { $t } from 'services/i18n';
 import ValidatedForm from 'components/shared/inputs/ValidatedForm.vue';
 
-import { IListMetadata, ITextMetadata, EInputType } from 'components/shared/inputs/index';
+import { IListMetadata, ITextMetadata, EInputType } from '../../../shared/inputs';
 
 @Component({
   components: {
