+++ app/components/page-components/Chatbot/windows/ChatbotQueuePreferencesWindow.vue.ts
@@ -5,7 +5,7 @@ import { $t } from 'services/i18n';
 import { metadata as metadataHelper } from 'components/widgets/inputs';
 import { IQueuePreferencesGeneralSettings } from 'services/chatbot';
 
-import { EInputType } from 'components/shared/inputs/index';
+import { EInputType } from '../../../shared/inputs';
 
 @Component({})
 export default class ChatbotQueuePreferencesWindow extends ChatbotWindowsBase {
