+++ app/components/page-components/Chatbot/windows/ChatbotSongRequestPreferencesWindow.vue.ts
@@ -5,7 +5,6 @@ import { ITab } from 'components/Tabs.vue';
 import { metadata as metadataHelper } from 'components/widgets/inputs';
 import { cloneDeep } from 'lodash';
 import { IMediaShareBan } from 'services/widgets/settings/media-share';
-import { EInputType } from 'components/shared/inputs/index';
 import ValidatedForm from 'components/shared/inputs/ValidatedForm.vue';
 
 import { ISongRequestData } from 'services/chatbot';
